Hotfix: Admin-Modus-Routing deaktiviert (Login-Loop) + int4-Decode-Fix
Der Admin-Modus-Default (#235) leitete den Admin auf die partner_response-
Präsentation (isAdmin:false). Das Frontend verhielt sich dann wie ein Partner
und schickte keinen Streamer-Override — internal-home verlangt für Admin-Level-
Sessions aber zwingend einen Override und antwortete mit 401 + loginUrl. Folge:
Endlos-Redirect zur Twitch-Anmeldung, Dashboard nicht ladbar.

auth-status liefert für Admin-Level wieder admin_response (adminEligible:false →
Frontend-Toggle bleibt versteckt). Endpoint + Frontend-Code bleiben dormant für
die saubere Neuauflage. Zusätzlich: partner_access.rs dekodierte
manual_partner_opt_out als Option<i64>, die Spalte ist aber INT4 → Decode brach
jeden Partner-Lookup; jetzt Option<i32>.

+++ rust/crates/tb-analytics/src/partner_access.rs
@@ -69,7 +69,9 @@ pub struct AccessState {
 struct PartnerRow {
     status: Option<String>,
     archived_at: Option<String>,
-    manual_partner_opt_out: Option<i64>,
+    // DB-Spalte ist INT4 (vgl. admin_streamers.rs); Option<i64> brach das Decode
+    // und ließ auth-status für jeden Partner-Lookup fail-open laufen.
+    manual_partner_opt_out: Option<i32>,
     technical_pause_reason: Option<String>,
 }
